LayoutTests/http/tests/security/contentSecurityPolicy/inline-style-attribute-blocked-expected.txt

CONSOLE MESSAGE: line 10: Refused to apply a stylesheet because its hash, its nonce, or 'unsafe-inline' does not appear in the style-src directive of the Content Security Policy.
PASS
